requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pandas>=2.2.0
numpy>=1.26.0
anthropic>=0.25.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
plotly>=5.19.0
seaborn>=0.13.0
matplotlib>=3.8.0
gunicorn>=21.2.0